#228 Further shorten filter names
