+++ chemclipse/plugins/org.eclipse.chemclipse.xxd.model/META-INF/MANIFEST.MF
@@ -49,7 +49,7 @@ Import-Package: com.google.common.collect,
 Service-Component: OSGI-INF/org.eclipse.chemclipse.xxd.model.filter.peaks.RemoveAllPeaksFilter.xml,
  OSGI-INF/org.eclipse.chemclipse.xxd.model.filter.peaks.DisableClassifiedPeaksFilter.xml,
  OSGI-INF/org.eclipse.chemclipse.xxd.model.filter.peaks.MinimumNumberOfIonsFilter.xml,
- OSGI-INF/org.eclipse.chemclipse.xxd.model.filter.peaks.PercentageAreaFilter.xml,
- OSGI-INF/org.eclipse.chemclipse.xxd.model.filter.peaks.PeakShapeFilter.xml,
- OSGI-INF/org.eclipse.chemclipse.xxd.model.filter.peaks.PeakAsymmetryFilter.xml,
- OSGI-INF/org.eclipse.chemclipse.xxd.model.filter.peaks.AreaFilter.xml
+ OSGI-INF/org.eclipse.chemclipse.xxd.model.filter.peaks.AreaFilter.xml,
+ OSGI-INF/org.eclipse.chemclipse.xxd.model.filter.peaks.AsymmetryFilter.xml,
+ OSGI-INF/org.eclipse.chemclipse.xxd.model.filter.peaks.ShapeFilter.xml,
+ OSGI-INF/org.eclipse.chemclipse.xxd.model.filter.peaks.AreaPercentFilter.xml
