+++ chemclipse/plugins/org.eclipse.chemclipse.xxd.model/OSGI-INF/org.eclipse.chemclipse.xxd.model.filter.peaks.AreaPercentFilter.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="org.eclipse.chemclipse.xxd.model.filter.peaks.PeakAsymmetryFilter">
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="org.eclipse.chemclipse.xxd.model.filter.peaks.AreaPercentFilter">
    <service>
       <provide interface="org.eclipse.chemclipse.model.filter.IPeakFilter"/>
       <provide interface="org.eclipse.chemclipse.processing.filter.Filter"/>
       <provide interface="org.eclipse.chemclipse.processing.Processor"/>
    </service>
-   <implementation class="org.eclipse.chemclipse.xxd.model.filter.peaks.PeakAsymmetryFilter"/>
+   <implementation class="org.eclipse.chemclipse.xxd.model.filter.peaks.AreaPercentFilter"/>
 </scr:component>
