+++ chemclipse/plugins/org.eclipse.chemclipse.xxd.model/src/org/eclipse/chemclipse/xxd/model/filter/peaks/AreaPercentFilterSettings.java
@@ -13,19 +13,23 @@ package org.eclipse.chemclipse.xxd.model.filter.peaks;
 
 import org.eclipse.chemclipse.support.settings.DoubleSettingsProperty;
 import org.eclipse.chemclipse.support.settings.EnumSelectionSettingProperty;
-import org.eclipse.chemclipse.xxd.model.support.PeakAsymmetryFilterSelectionCriterion;
+import org.eclipse.chemclipse.xxd.model.support.AreaFilterSelectionCriterion;
 import org.eclipse.chemclipse.xxd.model.support.ValueFilterTreatmentOption;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyDescription;
 
-public class PeakAsymmetryFilterSettings {
+public class AreaPercentFilterSettings {
 
-	@JsonProperty(value = "As:", defaultValue = "1.02")
-	@JsonPropertyDescription(value = "The peak asymmetry factor of a peak to be filtered accordingly.")
-	@DoubleSettingsProperty(minValue = 0.0d, maxValue = 10.0d)
-	// peak asymmetry factor (As) rating — As = 1.0-1.05 [excellent] - As = 1.2 [acceptable] - As >= 2 [unacceptable]
-	private double  peakAsymmetryFactor = 1.02d;
+	@JsonProperty(value = "Minimum area:")
+	@JsonPropertyDescription(value = "The minimum percentage area value of a peak to be filtered accordingly.")
+	@DoubleSettingsProperty(minValue = 0.0d, maxValue = 100.0d)
+	private double minimumPercentageAreaValue = 1.0d;
+
+	@JsonProperty(value = "Maximum area:")
+	@JsonPropertyDescription(value = "The maximum percentage area value of a peak to be filtered accordingly.")
+	@DoubleSettingsProperty(minValue = 0.0d, maxValue = 100.0d)
+	private double maximumPercentageAreaValue = 10.0d;
 
 	@JsonProperty(value = "Peak Treatment Option:")
 	@EnumSelectionSettingProperty
@@ -33,34 +37,45 @@ public class PeakAsymmetryFilterSettings {
 
 	@JsonProperty(value = "Peak Selection Criterion:")
 	@EnumSelectionSettingProperty
-	private PeakAsymmetryFilterSelectionCriterion filterSelectionCriterion = PeakAsymmetryFilterSelectionCriterion.ASYMMETRY_FACTOR_GREATER_THAN_LIMIT;
+	private AreaFilterSelectionCriterion filterSelectionCriterion = AreaFilterSelectionCriterion.AREA_LESS_THAN_MINIMUM;
 
-	public double getPeakAsymmetryFactor() {
-		return peakAsymmetryFactor;
+	public double getMinimumPercentageAreaValue() {
+		
+		return minimumPercentageAreaValue;
 	}
 
-	public void setPeakAsymmetryFactor(double peakAsymmetryFactor) {
+	public void setMinimumPercentageAreaValue(double minimumPercentageAreaValue) {
+		
+		this.minimumPercentageAreaValue = minimumPercentageAreaValue;
+	}
 
-		this.peakAsymmetryFactor = peakAsymmetryFactor;
+	public double getMaximumPercentageAreaValue() {
+		
+		return maximumPercentageAreaValue;
+	}
+
+	public void setMaximumPercentageAreaValue(double maximumPercentageAreaValue) {
+		
+		this.maximumPercentageAreaValue = maximumPercentageAreaValue;
 	}
 
 	public ValueFilterTreatmentOption getFilterTreatmentOption() {
-
+		
 		return filterTreatmentOption;
 	}
 
 	public void setFilterTreatmentOption(ValueFilterTreatmentOption filterTreatmentOption) {
-
+		
 		this.filterTreatmentOption = filterTreatmentOption;
 	}
 
-	public PeakAsymmetryFilterSelectionCriterion getFilterSelectionCriterion() {
-
+	public AreaFilterSelectionCriterion getFilterSelectionCriterion() {
+		
 		return filterSelectionCriterion;
 	}
 
-	public void setFilterSelectionCriterion(PeakAsymmetryFilterSelectionCriterion filterSelectionCriterion) {
-
+	public void setFilterSelectionCriterion(AreaFilterSelectionCriterion filterSelectionCriterion) {
+		
 		this.filterSelectionCriterion = filterSelectionCriterion;
 	}
 }
