+++ chemclipse/plugins/org.eclipse.chemclipse.xxd.model/src/org/eclipse/chemclipse/xxd/model/filter/peaks/AsymmetryFilter.java
@@ -26,24 +26,24 @@ import org.eclipse.core.runtime.SubMonitor;
 import org.osgi.service.component.annotations.Component;
 
 @Component(service = { IPeakFilter.class, Filter.class, Processor.class })
-public class PercentageAreaFilter implements IPeakFilter<PercentageAreaFilterSettings> {
+public class AsymmetryFilter implements IPeakFilter<AsymmetryFilterSettings> {
 
 	@Override
 	public String getName() {
 
-		return "Percentage Area Filter";
+		return "Peak Asymmetry Filter";
 	}
 
 	@Override
 	public String getDescription() {
 
-		return "Filter peaks by percentage area values";
+		return "Filter peaks by asymmetrical peak shape";
 	}
 
 	@Override
-	public Class<PercentageAreaFilterSettings> getConfigClass() {
+	public Class<AsymmetryFilterSettings> getConfigClass() {
 
-		return PercentageAreaFilterSettings.class;
+		return AsymmetryFilterSettings.class;
 	}
 
 	@Override
@@ -53,72 +53,46 @@ public class PercentageAreaFilter implements IPeakFilter<PercentageAreaFilterSet
 	}
 
 	@Override
-	public <X extends IPeak> void filterIPeaks(CRUDListener<X, IPeakModel> listener, PercentageAreaFilterSettings configuration, MessageConsumer messageConsumer, IProgressMonitor monitor) throws IllegalArgumentException {
+	public <X extends IPeak> void filterIPeaks(CRUDListener<X, IPeakModel> listener, AsymmetryFilterSettings configuration, MessageConsumer messageConsumer, IProgressMonitor monitor) throws IllegalArgumentException {
 
 		Collection<X> read = listener.read();
 		if(configuration == null) {
 			configuration = createConfiguration(read);
 		}
 		SubMonitor subMonitor = SubMonitor.convert(monitor, read.size());
-		double areaSum = calculateAreaSum(read);
 
 		for(X peak : read) {
-			double compareAreaValue = calculatePercentageAreaCompareValue(peak, areaSum);
-			applySelectedOptions(compareAreaValue, configuration, listener, peak);
+			applySelectedOptions(configuration, listener, peak);
 			subMonitor.worked(1);
 		}
 	}
 
-	private static <X extends IPeak> double calculateAreaSum(Collection<X> read) {
-
-		double areaSum = 0;
-		for(X peak : read) {
-			areaSum = areaSum + peak.getIntegratedArea();
-		}
-		return areaSum;
-	}
-
-	private static <X extends IPeak> double calculatePercentageAreaCompareValue(X peak, double areaSum) {
-
-		return (100 / areaSum) * peak.getIntegratedArea();
-	}
-
-	private static <X extends IPeak> void applySelectedOptions(double peakValue, PercentageAreaFilterSettings configuration, CRUDListener<X, IPeakModel> listener, X peak) {
+	private static <X extends IPeak> void applySelectedOptions(AsymmetryFilterSettings configuration, CRUDListener<X, IPeakModel> listener, X peak) {
 
 		boolean keepFlag = false;
 		if(configuration.getFilterTreatmentOption()==ValueFilterTreatmentOption.KEEP_PEAK) {
 			keepFlag = true;
 		}
-		switch (configuration.getFilterSelectionCriterion()) {
-		case AREA_LESS_THAN_MINIMUM:
+		double  peakAsymmetryFactor = peak.getPeakModel().getTailing()/peak.getPeakModel().getLeading();
+		switch (configuration.getFilterSelectionCriterion()){
+		case ASYMMETRY_FACTOR_GREATER_THAN_LIMIT:
 			if(keepFlag) {
-				if(Double.compare(peakValue, configuration.getMinimumPercentageAreaValue())>0) {
+				if(Double.compare(peakAsymmetryFactor, configuration.getPeakAsymmetryFactor())<0) {
 					processPeak(listener, configuration, peak);
 				}
 			} else {
-				if(Double.compare(peakValue, configuration.getMinimumPercentageAreaValue())<0) {
+				if(Double.compare(peakAsymmetryFactor, configuration.getPeakAsymmetryFactor())>0) {
 					processPeak(listener, configuration, peak);
 				}
 			}
 			break;
-		case AREA_GREATER_THAN_MAXIMUM:
+		case ASYMMETRY_FACTOR_SMALLER_THAN_LIMIT:
 			if(keepFlag) {
-				if(Double.compare(peakValue, configuration.getMaximumPercentageAreaValue())<0) {
+				if(Double.compare(peakAsymmetryFactor, configuration.getPeakAsymmetryFactor())>0) {
 					processPeak(listener, configuration, peak);
 				}
 			} else {
-				if(Double.compare(peakValue, configuration.getMaximumPercentageAreaValue())>0) {
-					processPeak(listener, configuration, peak);
-				}
-			}
-			break;
-		case AREA_NOT_WITHIN_RANGE:
-			if(keepFlag) {
-				if(!checkRange(peakValue, configuration)) {
-					processPeak(listener, configuration, peak);
-				}
-			} else {
-				if(checkRange(peakValue, configuration)) {
+				if(Double.compare(peakAsymmetryFactor, configuration.getPeakAsymmetryFactor())<0) {
 					processPeak(listener, configuration, peak);
 				}
 			}
@@ -128,14 +102,9 @@ public class PercentageAreaFilter implements IPeakFilter<PercentageAreaFilterSet
 		}
 	}
 
-	private static boolean checkRange(double peakValue, PercentageAreaFilterSettings configuration) {
-		
-		return Double.compare(peakValue, configuration.getMinimumPercentageAreaValue())<0 || Double.compare(peakValue, configuration.getMaximumPercentageAreaValue())>0;
-	}
+	private static <X extends IPeak> void processPeak(CRUDListener<X, IPeakModel> listener, AsymmetryFilterSettings configuration, X peak) {
 
-	private static <X extends IPeak> void processPeak(CRUDListener<X, IPeakModel> listener, PercentageAreaFilterSettings localSettings, X peak) {
-
-		switch (localSettings.getFilterTreatmentOption()) {
+		switch (configuration.getFilterTreatmentOption()) {
 		case ENABLE_PEAK:
 			peak.setActiveForAnalysis(true);
 			listener.updated(peak);
